modif controller book

+++ backend/controllers/book.js
@@ -1,4 +1,5 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 
 exports.createBook = (req, res, next) => {
@@ -16,13 +17,6 @@ exports.createBook = (req, res, next) => {
     .catch(error => { res.status(400).json( { error })})
  };
 
-
-// exports.modifyBook = (req, res, next) => {
-// Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-//     .then(() => res.status(200).json({ message: 'Livre modifié !'}))
-//     .catch(error => res.status(400).json({ error }));
-// }
-
 exports.modifyBook = (req, res, next) => {
 const bookObject = req.file ? {
     ...JSON.parse(req.body.book),
@@ -36,7 +30,7 @@ Book.findOne({_id: req.params.id})
             res.status(401).json({ message : 'Not authorized'});
         } else {
             Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
-            .then(() => res.status(200).json({message : 'Objet modifié!'}))
+            .then(() => res.status(200).json({message : 'Livre modifié!'}))
             .catch(error => res.status(401).json({ error }));
         }
     })
@@ -45,12 +39,24 @@ Book.findOne({_id: req.params.id})
     });
 };
 
-
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Livre supprimé !'}))
-        .catch(error => res.status(400).json({ error }));
-}
+    Book.findOne({ _id: req.params.id})
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Livre supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+ };
 
 
 exports.getOneBook = (req, res, next) => {
